wallet: generalize SQL wallet helpers
Move wallet parameter construction and address derivation out of the
SQLite backend. Give the shared helpers database-neutral names so
PostgreSQL can use the same wallet creation behavior.
